% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MATpca.R
\name{MATpca}
\alias{MATpca}
\title{MATpca: Analisis de Componentes Principales}
\usage{
MATpca(data, ...)
}
\arguments{
\item{data}{Dataframe con los datos.}

\item{...}{Variables especificas (sin comillas) a incluir en el analisis.}
}
\value{
Lista con el objeto PCA (prcomp) y una lista 'info' con tablas (kable) y graficos.
}
\description{
Realiza un Analisis de Componentes Principales (PCA) sobre variables metricas.
Devuelve tablas de resumen y graficos de autovalores y varianza acumulada.
}
